@@ -1,12 +1,14 @@
 import { CommonService } from './services/common.service';
 import { Component } from '@angular/core';
-import { Router } from '@angular/router';
+import { Router, RouterOutlet } from '@angular/router';
+import { HomeComponent } from './home/home.component';
 
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.scss'],
-    standalone: false
+    standalone: true,
+    imports: [HomeComponent, RouterOutlet]
 })
 export class AppComponent {
   title = 'streaming';
