@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-home',
- standalone: false,
+ standalone: true,
   templateUrl: './home.component.html',
   styleUrl: './home.component.scss'
 })
